# src/server/api/swagger.json
{
  "swagger": "2.0",
  "produces": [
    "application/json"
  ],
  "info": {
    "title": "conversationdiscovery",
    "description": "An API",
    "version": "1.0"
  },
  "basePath": "/conversationdiscovery",
  "schemes": [
    "http",
    "https"
  ],
  "parameters": {
    "clientId": {
      "name": "clientId",
      "in": "path",
      "description": "A client id, e.g. \"247inc\".",
      "required": true,
      "type": "string"
    },
    "applicationId": {
      "name": "applicationId",
      "in": "path",
      "description": "An app id, e.g. \"ReferenceBot\".",
      "required": true,
      "type": "string"
    },
...
    "accountId": {
      "name": "accountId",
      "in": "path",
      "description": "An account id, e.g. \"ReferenceBot\".",
      "required": true,
      "type": "string"
    },
...
    "runId": {
      "name": "runId",
      "in": "path",
      "description": "Run id of a run",
      "required": true,
      "type": "string",
      "format": "uuid"
    },
...
    "projectId": {
      "name": "projectId",
      "in": "path",
      "description": "id of the project",
      "required": true,
      "type": "string"
    },
...
  },
  "responses": {
    "BadRequest": {
      "description": "The request was invalid. For example, the value of the initial parameter was invalid",
      "schema": {
        "type": "string"
      }
    },
    "NotFound": {
      "description": "The URI requested is invalid or the resource requested, such as a user, does not exist. Do not retry with the same input",
      "schema": {
        "type": "string"
      }
    },
    "Unauthorized": {
      "description": "Missing or incorrect authentication credentials. This may also returned in other undefined circumstances.Do not retry with the same input",
      "schema": {
        "type": "string"
      }
    },
    "Forbidden": {
      "description": "Forbidden. The request is understood, but it has been refused or access is not allowed. An accompanying error message will explain further. Do not retry with the same input",
      "schema": {
        "type": "string"
      }
    },
...
    "InternalServerError": {
      "description": "Internal server error",
      "schema": {
        "type": "string"
      }
    },
    "Conflict": {
      "description": "The request could not be completed due to a conflict with the current state of the resource.Do not retry with the same input.",
      "schema": {
        "type": "string"
      }
    }
  },
  "paths": {
...
    "/clients/{clientId}/accounts/{accountId}/applications/{applicationId}/projects/{projectId}": {
      "patch": {
        "summary": "Updates the project name and/or project description for a given client, app, account and projectId",
        "x-swagger-router-controller": "projects",
        "operationId": "updateProject",
        "tags": [
          "project"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/clientId"
          },
          {
            "$ref": "#/parameters/applicationId"
          },
          {
            "$ref": "#/parameters/accountId"
          },
          {
            "$ref": "#/parameters/projectId"
          },
          {
            "in": "body",
            "name": "Project name",
            "description": "The project to update",
            "schema": {
              "type": "object",
              "properties": {
                "projectName": {
                  "type": "string",
                  "minLength": 4,
                  "maxLength": 50
                },
                "projectDescription": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 150
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated a project",
            "schema": {
              "$ref": "#/definitions/Project"
            }
          },
          "400": {
            "$ref": "#/responses/BadRequest"
          },
          "401": {
            "$ref": "#/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/responses/Forbidden"
          },
          "404": {
            "$ref": "#/responses/NotFound"
          },
          "409": {
            "$ref": "#/responses/Conflict"
          }
        }
      },
      "delete": {
        "summary": "Disables the project for a given client, app, account and projectId",
        "x-swagger-router-controller": "projects",
        "operationId": "deleteProject",
        "tags": [
          "project"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/clientId"
          },
          {
            "$ref": "#/parameters/applicationId"
          },
          {
            "$ref": "#/parameters/accountId"
          },
          {
            "$ref": "#/parameters/projectId"
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted a project",
            "schema": {
              "$ref": "#/definitions/Project"
            }
          },
          "401": {
            "$ref": "#/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/responses/Forbidden"
          },
          "404": {
            "$ref": "#/responses/NotFound"
          },
          "500": {
            "$ref": "#/responses/InternalServerError"
          }
        }
      },
      "get": {
        "summary": "Get the project for a given app",
        "x-swagger-router-controller": "projects",
        "operationId": "getProject",
        "tags": [
          "project"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/clientId"
          },
          {
            "$ref": "#/parameters/applicationId"
          },
          {
            "$ref": "#/parameters/accountId"
          },
          {
            "$ref": "#/parameters/projectId"
          }
        ],
        "responses": {
          "200": {
            "description": "Retrieved Project",
            "schema": {
              "$ref": "#/definitions/Project"
            }
          },
          "500": {
            "$ref": "#/responses/InternalServerError"
          },
          "401": {
            "$ref": "#/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/responses/Forbidden"
          },
          "404": {
            "$ref": "#/responses/NotFound"
          }
        }
      }
    },
    "/clients/{clientId}/accounts/{accountId}/applications/{applicationId}/projects/{projectId}/runs": {
      "post": {
        "summary": "Creates a new run for a given project",
        "x-swagger-router-controller": "runs",
        "operationId": "createRun",
        "tags": [
          "run"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/clientId"
          },
          {
            "$ref": "#/parameters/applicationId"
          },
          {
            "$ref": "#/parameters/accountId"
          },
          {
            "$ref": "#/parameters/projectId"
          },
          {
            "in": "body",
            "name": "data",
            "required": true,
            "description": "The run to create",
            "schema": {
              "$ref": "#/definitions/RunCreationObject"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Created a new run",
            "schema": {
              "$ref": "#/definitions/RunAfterCreation"
            }
          },
          "400": {
            "$ref": "#/responses/BadRequest"
          },
          "401": {
            "$ref": "#/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/responses/NotFound"
          }
        }
      },
      "get": {
        "summary": "Get all the runs for a given project",
        "x-swagger-router-controller": "runs",
        "operationId": "getAllRuns",
        "tags": [
          "run"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/clientId"
          },
          {
            "$ref": "#/parameters/applicationId"
          },
          {
            "$ref": "#/parameters/accountId"
          },
          {
            "$ref": "#/parameters/projectId"
          }
        ],
        "responses": {
          "200": {
            "description": "All the runs for a given project",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Run"
              }
            }
          },
          "404": {
            "$ref": "#/responses/NotFound"
          },
          "500": {
            "$ref": "#/responses/InternalServerError"
          }
        }
      }
    },
    "/clients/{clientId}/accounts/{accountId}/applications/{applicationId}/projects/{projectId}/runs/{runId}": {
      "get": {
        "summary": "Get given run for the given project",
        "x-swagger-router-controller": "runs",
        "operationId": "getRun",
        "tags": [
          "run"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/clientId"
          },
          {
            "$ref": "#/parameters/applicationId"
          },
          {
            "$ref": "#/parameters/accountId"
          },
          {
            "$ref": "#/parameters/projectId"
          },
          {
            "$ref": "#/parameters/runId"
          }
        ],
        "responses": {
          "200": {
            "description": "The run object for the given runId",
            "schema": {
              "$ref": "#/definitions/Run"
            }
          },
          "500": {
            "$ref": "#/responses/InternalServerError"
          },
          "404": {
            "$ref": "#/responses/NotFound"
          }
        }
      },
      "patch": {
        "summary": "Update the run name/description for a given client, app, account, project and run",
        "x-swagger-router-controller": "runs",
        "operationId": "updateRun",
        "tags": [
          "run"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/clientId"
          },
          {
            "$ref": "#/parameters/applicationId"
          },
          {
            "$ref": "#/parameters/accountId"
          },
          {
            "$ref": "#/parameters/projectId"
          },
          {
            "$ref": "#/parameters/runId"
          },
          {
            "in": "body",
            "name": "Run name",
            "description": "The run to update",
            "schema": {
              "type": "object",
              "properties": {
                "runName": {
                  "type": "string",
                  "minLength": 4,
                  "maxLength": 50,
                  "description": "The name of the run"
                },
                "runDescription": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 150,
                  "description": "The run description"
                },
                "starred": {
                  "description": "Describes if the run is starred",
                  "type": "integer"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated the run",
            "schema": {
              "$ref": "#/definitions/Run"
            }
          },
          "400": {
            "$ref": "#/responses/BadRequest"
          }
        }
      },
      "delete": {
        "summary": "Deletes a run",
        "x-swagger-router-controller": "runs",
        "operationId": "deleteRun",
        "tags": [
          "run"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/clientId"
          },
          {
            "$ref": "#/parameters/applicationId"
          },
          {
            "$ref": "#/parameters/accountId"
          },
          {
            "$ref": "#/parameters/projectId"
          },
          {
            "$ref": "#/parameters/runId"
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted a run",
            "schema": {
              "$ref": "#/definitions/RunAfterDeletion"
            }
          },
          "401": {
            "$ref": "#/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/responses/NotFound"
          },
          "500": {
            "$ref": "#/responses/InternalServerError"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "RunAfterDeletion": {
      "type": "object",
      "properties": {
        "runId": {
          "description": "The run id of the run",
          "type": "string",
          "format": "uuid"
        }
      }
    },
    "RunCreationObject": {
      "description": "The app publish params",
      "required": [
        "runName",
        "numOfClusters",
        "numOfTurns"
      ],
      "properties": {
        "runName": {
          "type": "string",
          "minLength": 4,
          "maxLength": 50,
          "description": "The name of the run"
        },
        "runDescription": {
          "type": "string",
          "minLength": 1,
          "maxLength": 150,
          "description": "The run description"
        },
        "numOfClusters": {
          "type": "integer",
          "minimum": 2,
          "maximum": 900,
          "description": "The number of clusters"
        },
        "numOfTurns": {
          "type": "integer",
          "minimum": 1,
          "maximum": 5,
          "description": "The number of client turns"
        },
        "stopWords": {
          "type": "string",
          "description": "The stopwords"
        }
      }
    },
...
    "RunAfterCreation": {
      "description": "Project Object just after creating a project",
      "type": "object",
      "properties": {
        "applicationId": {
          "description": "The app id of the Project",
          "type": "string"
        },
        "clientId": {
          "description": "The client id of the Project",
          "type": "string"
        },
        "accountId": {
          "description": "The account id of the Project",
          "type": "string"
        },
        "projectId": {
          "description": "The id of the project",
          "type": "string"
        },
        "runName": {
          "description": "The name of the run",
          "type": "string"
        },
        "runId": {
          "description": "The id of the run",
          "type": "string"
        },
        "runDescription": {
          "description": "The description of the run",
          "type": "string"
        },
        "numOfClusters": {
          "description": "The description of the run",
          "type": "integer"
        },
        "numOfTurns": {
          "description": "The number of client turns",
          "type": "integer"
        },
        "stopWords": {
          "description": "The stop words",
          "type": "string"
        },
        "created": {
          "description": "The UTC timestamp of time when the run was created",
          "type": "string"
        },
        "createdBy": {
          "description": "The id of the person who created the run",
          "type": "string"
        },
        "modified": {
          "description": "The UTC timestamp of time when the run was modified",
          "type": "string"
        },
        "modifiedBy": {
          "description": "The id of the person who modified the run",
          "type": "string"
        },
        "runStatus": {
          "description": "The status of the run",
          "type": "string"
        },
        "runStatusDescription": {
          "description": "The status description of the run",
          "type": "string"
        },
        "starred": {
          "description": "Check to determine if the run is starred",
          "type": "integer"
        }
      }
    },
    "Project": {
      "description": "Project Object",
      "type": "object",
      "properties": {
        "applicationId": {
          "description": "The app id of the Project",
          "type": "string"
        },
        "clientId": {
          "description": "The client id of the Project",
          "type": "string"
        },
        "accountId": {
          "description": "The account id of the Project",
          "type": "string"
        },
        "projectName": {
          "description": "The name of the project",
          "type": "string"
        },
        "datasetName": {
          "description": "The name of the project",
          "type": "string"
        },
        "projectDescription": {
          "description": "The description of the project",
          "type": "string"
        },
        "modified": {
          "description": "The UTC timestamp when the project was modified",
          "type": "number"
        },
        "modifiedBy": {
          "description": "The id of the person who modified the project",
          "type": "string"
        },
        "created": {
          "description": "The UTC timestamp when the project was created",
          "type": "number"
        },
        "createdBy": {
          "description": "The id of the person who created the project",
          "type": "string"
        },
        "projectStatus": {
          "description": "The status of the project",
          "type": "string"
        },
        "projectStatusDescription": {
          "description": "The description of the status of the project",
          "type": "string"
        }
      }
    },
...
  }
}